mejoras update

+++ app/controllers/profiles_controller.rb
@@ -62,31 +62,22 @@ class ProfilesController < ApplicationController
   end
 
   def update
-    user = current_user
-
     # Recuperacion de imagen desde params
     image_params_cover = params[:cover_image] if params[:cover_image].present?
     image_params_profile = params[:profile_image] if params[:profile_image].present?
 
     #### CREATE UPDATE IMAGES ####
-    update_image(user, image_params_cover, 'cover') if image_params_cover.present?
-    update_image(user, image_params_profile, 'profile') if image_params_profile.present?
+    update_image(current_user, image_params_cover, 'cover') if image_params_cover.present?
+    update_image(current_user, image_params_profile, 'profile') if image_params_profile.present?
 
     #### DESTROY IMAGES ####
-    if params[:destroy_cover_image].present? && user.cover_image.present?
-      user.cover_image.destroy
-      user.update(cover_image: nil)
-    end
+    current_user.cover_image.destroy if params[:destroy_cover_image].present? && current_user.cover_image.present?
+    current_user.profile_image.destroy if params[:destroy_profile_image].present? && current_user.profile_image.present?
 
-    if params[:destroy_profile_image].present? && user.profile_image.present?
-      user.profile_image.destroy
-      user.update(profile_image: nil)
-    end
-
-    if user.update(user_params_update)
-      render json: user.as_json(USER_TO_JSON_UPDATE), status: :ok
+    if current_user.update(user_params_update)
+      render json: current_user.reload.as_json(USER_TO_JSON_UPDATE), status: :ok
     else
-      render json: { errors: user.errors.details }, status: :unprocessable_entity
+      render json: { errors: current_user.errors.details }, status: :unprocessable_entity
     end
   end
